pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pempy"
version = "0.2.0"
description = "PEM electrolysis cell control for tritium enrichment"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.8"
dependencies = [
    "minimalmodbus>=2.0",
    "packaging",
    "pyserial>=3.0",
    "RPi.GPIO",
]

[project.scripts]
pemcell = "pempy.cli.pemcell:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]
